flask_app/requirements.txt
mysql-connector-python
flask
marshmallow-dataclass